04-task/agentic-app/backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "codeWithHak", email = "huzairahmedkhan@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "openai-agents>=0.6.1",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "supabase>=2.24.0",
]

[project.scripts]
app = "app.main:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
